Introduces types ItemType and ItemQuantity, and Deal as entities
At this point, it's purely for additional clarity for what the
primitives (string and uint) represent. However it will allow for the
extension/changing of these types later on, hopefully making our design
more flexible.

The deal entity and its methods are added here as well.

+++ checkout/get_total_price.go
@@ -4,7 +4,7 @@ package checkout
 func (r Repository) GetTotalPrice() uint {
 	var totalPrice uint
 	for name, quantity := range r.basket {
-		totalPrice += quantity * r.priceChecker.GetItemPrice(name)
+		totalPrice += uint(quantity) * r.priceChecker.GetItemPrice(name)
 	}
 	return totalPrice
 }
